break up files, cleanup
